remove GetAllGenesisValidatorByLaunchID method
